Tidy-up, fix some issues
