+++ src/Bicep.Decompiler/ArmHelpers/TemplateHelpers.cs
@@ -103,6 +103,15 @@ namespace Bicep.Decompiler.ArmHelpers
                 
                 var (childType, childName, _) = ParseResource(childResourceObject);
 
+                if (GetProperty(resource, "copy") is {} copyProperty)
+                {
+                    childResourceObject["copy"] = copyProperty.Value;
+                }
+                if (GetProperty(resource, "condition") is {} conditionProperty)
+                {
+                    childResourceObject["condition"] = conditionProperty.Value;
+                }
+
                 // child may sometimes be specified using the fully-qualified type and name
                 if (!StringComparer.OrdinalIgnoreCase.Equals(parentType.Split("/")[0], childType.Split("/")[0]))
                 {
@@ -117,197 +126,12 @@ namespace Bicep.Decompiler.ArmHelpers
                 }
             }
         }
-
-        public static class LanguageExpressionRewriter
-        {
-            public static string Rewrite(string value, Func<LanguageExpression, LanguageExpression> rewriteFunc)
-            {
-                LanguageExpression expression;
-                if (ExpressionsEngine.IsLanguageExpression(value))
-                {
-                    expression = ExpressionsEngine.ParseLanguageExpression(value);
-                }
-                else
-                {
-                    expression = new JTokenExpression(value);
-                }
-
-                var (hasChanges, rewritten) = RewriteInternal(expression, rewriteFunc);
-                
-                return hasChanges ? ExpressionsEngine.SerializeExpression(rewritten, new ExpressionSerializerSettings { IncludeOuterSquareBrackets = true }) : value;
-            }
-
-            private static (bool hasChanges, LanguageExpression[] output) RewriteInternal(LanguageExpression[] input, Func<LanguageExpression, LanguageExpression> rewriteFunc)
-            {
-                var output = new LanguageExpression[input.Length];
-
-                var hasChanges = false;
-                for (var i = 0; i < input.Length; i++)
-                {
-                    var (entryHasChanges, entryOutput) = RewriteInternal(input[i], rewriteFunc);
-
-                    hasChanges |= entryHasChanges;
-                    output[i] = entryOutput;
-                }
-
-                return (hasChanges, output);
-            }
-
-            private static (bool hasChanges, LanguageExpression output) RewriteInternal(LanguageExpression input, Func<LanguageExpression, LanguageExpression> rewriteFunc)
-            {
-                switch (input)
-                {
-                    case FunctionExpression function:
-                    {
-                        var hasChanges = false;
-                        
-                        var newParameters = function.Parameters;
-                        if (function.Parameters is not null)
-                        {
-                            var (hasChangesInner, output) = RewriteInternal(function.Parameters, rewriteFunc);
-                            hasChanges |= hasChangesInner;
-                            newParameters = output;
-                        }
-
-                        var newProperties = function.Properties;
-                        if (function.Properties is not null)
-                        {
-                            var (hasChangesInner, output) = RewriteInternal(function.Properties, rewriteFunc);
-                            hasChanges |= hasChangesInner;
-                            newProperties = output;
-                        }
-
-                        if (hasChanges)
-                        {
-                            function = new FunctionExpression(function.Function, newParameters, newProperties);
-                        }
-
-                        var rewritten = rewriteFunc(function);
-                        hasChanges |= !object.ReferenceEquals(function, rewritten);
-
-                        return (hasChanges, rewritten);
-                    }
-                    case JTokenExpression jtoken:
-                    {
-                        var rewritten = rewriteFunc(jtoken);
-                        var hasChanges = !object.ReferenceEquals(jtoken, rewritten);
-
-                        return (hasChanges, rewritten);
-                    }
-                    default:
-                        throw new NotImplementedException($"Unrecognized expression type {input.GetType()}");
-                }
-            }
-        }
-
-        private static JObject RewriteExpressions(JObject input, Func<LanguageExpression, LanguageExpression> rewriteFunc)
-        {
-            if (input.DeepClone() is not JObject clonedInput)
-            {
-                throw new InvalidOperationException($"Failed to clone input");
-            }
-
-            input = clonedInput;
-
-            JsonUtility.WalkJsonRecursive(
-                input,
-                objectAction: @object => {
-                    foreach (var property in @object.Properties())
-                    {
-                        var newName = LanguageExpressionRewriter.Rewrite(property.Name, rewriteFunc);
-                        if (newName != property.Name)
-                        {
-                            property.AddBeforeSelf(new JProperty(newName, property.Value));
-                            property.Remove();
-                        }
-                    }
-                },
-                propertyAction: property => {
-                    if (property.Value is null)
-                    {
-                        return;
-                    }
-
-                    if (property.Value.Type == JTokenType.String && property.Value.ToObject<string>() is string value)
-                    {
-                        var newValue = LanguageExpressionRewriter.Rewrite(value, rewriteFunc);
-                        if (newValue != value)
-                        {
-                            property.Value.Replace(newValue);
-                        }
-                    }
-                    else if (property.Value.Type == JTokenType.Array)
-                    {
-                        foreach (var child in property.Value.Children())
-                        {
-                            if (child.Type == JTokenType.String && child.ToObject<string>() is string childValue)
-                            {
-                                var newValue = LanguageExpressionRewriter.Rewrite(childValue, rewriteFunc);
-                                if (newValue != childValue)
-                                {
-                                    child.Replace(newValue);
-                                }
-                            }
-                        }
-                    }
-                });
-
-            return input;
-        }
-
-        private static void VisitExpressions(JToken? input, Action<LanguageExpression> visitFunc)
-        {
-            if (input is null)
-            {
-                return;
-            }
-
-            var visitor = new LanguageExpressionVisitor
-            {
-                OnFunctionExpression = visitFunc,
-                OnJTokenExpression = visitFunc,
-            };
-
-            JsonUtility.WalkJsonRecursive(
-                input,
-                objectAction: @object => {
-                    foreach (var property in @object.Properties())
-                    {
-                        if (ExpressionsEngine.IsLanguageExpression(property.Name))
-                        {
-                            var expression = ExpressionsEngine.ParseLanguageExpression(property.Name);
-                            expression.Accept(visitor);
-                        }
-                        else
-                        {
-                            var expression = new JTokenExpression(property.Name);
-                            expression.Accept(visitor);
-                        }
-                    }
-                },
-                tokenAction: token => {
-                    if (token.Type == JTokenType.String && token.ToObject<string>() is string value)
-                    {
-                        if (ExpressionsEngine.IsLanguageExpression(value))
-                        {
-                            var expression = ExpressionsEngine.ParseLanguageExpression(value);
-                            expression.Accept(visitor);
-                        }
-                        else
-                        {
-                            var expression = new JTokenExpression(value);
-                            expression.Accept(visitor);
-                        }
-                    }
-                });
-        }
-
         public static (JObject template, IReadOnlyDictionary<string, FunctionExpression> parameters) ConvertNestedTemplateInnerToOuter(JObject template)
         {
             // this is useful to avoid naming clashes - we should prioritize names that have come from the parent template
             var paramsAccessed = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
 
-            VisitExpressions(template, expression => 
+            ExpressionHelpers.VisitExpressions(template, expression => 
             {
                 if (expression is not FunctionExpression function || !IsFunction(function, "parameters"))
                 {
@@ -353,7 +177,7 @@ namespace Bicep.Decompiler.ArmHelpers
             }
 
             // process references first
-            template = RewriteExpressions(template, expression =>
+            template = ExpressionHelpers.RewriteExpressions(template, expression =>
             {
                 if (expression is not FunctionExpression function || !IsFunction(function, "reference"))
                 {
@@ -370,7 +194,7 @@ namespace Bicep.Decompiler.ArmHelpers
             });
 
             // process resourceIds
-            template = RewriteExpressions(template, expression => 
+            template = ExpressionHelpers.RewriteExpressions(template, expression => 
             {
                 if (expression is not FunctionExpression function || !IsFunction(function, "resourceId"))
                 {
@@ -381,7 +205,7 @@ namespace Bicep.Decompiler.ArmHelpers
             });
 
             // process variables
-            template = RewriteExpressions(template, expression => 
+            template = ExpressionHelpers.RewriteExpressions(template, expression => 
             {
                 if (expression is not FunctionExpression function || !IsFunction(function, "variables"))
                 {
@@ -392,7 +216,7 @@ namespace Bicep.Decompiler.ArmHelpers
             });
 
             // add parameters to lookup
-            VisitExpressions(template, expression => 
+            ExpressionHelpers.VisitExpressions(template, expression => 
             {
                 if (expression is not FunctionExpression function || !IsFunction(function, "parameters"))
                 {
